fix(aggregate-accessibility): Fix the selector for comparison aggregate accessibility.
fixes #753
