+++ lib/selectors/comparison-aggregate-accessibility.js
@@ -1,13 +1,14 @@
 // @flow
 import {createSelector} from 'reselect'
 
-import {getAggregateAccessibility} from './aggregate-accessibility'
+import {activeOpportunityDatasetGrid} from '../modules/opportunity-datasets/selectors'
+import getAggregateAccessibility from '../utils/aggregate-accessibility'
 
 /** Aggregate accessibility for the comparison regional analysis */
 export default createSelector(
   state => state.analysis.regional.comparisonGrid,
   // aggregation area and weights don't vary between base and comparison
   state => state.analysis.regional.aggregationArea,
-  state => state.analysis.regional.aggregationWeights,
+  activeOpportunityDatasetGrid,
   getAggregateAccessibility
 )
