+++ lib/utils/__tests__/aggregate-accessibility.js
@@ -7,12 +7,12 @@
 
 import range from 'lodash/range'
 
-import {getAggregateAccessibility} from '../aggregate-accessibility'
+import getAggregateAccessibility from '../aggregate-accessibility'
 
 const contains = (width, height) => (x, y) =>
   x >= 0 && y >= 0 && x < width && y < height
 
-describe('Selectors > aggregation accessibility', () => {
+describe('utils > aggregation accessibility', () => {
   it('should work', () => {
     // if we think of these as population, the two right cells of the top row have population of 100 and 25
     const population = {
